backend/requirements.txt
# fastapi[all]==0.104.1
# uvicorn[standard]==0.24.0
# sqlalchemy==2.0.23
# pymysql==1.1.0
# python-dotenv==1.0.0
# passlib[bcrypt]==1.7.4
# python-jose[cryptography]==3.3.0
# python-multipart==0.0.6
# httpx==0.25.0   # for any async HTTP calls



fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.26  # updated to support Python 3.13
pymysql==1.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.0
argon2-cffi